feat: Add ServerStart class for starting the network server
This commit adds a new file, ServerStart.java, which contains the implementation for starting the network server. The ServerStart class is annotated with @CommandLine.Command and implements the Runnable interface. It includes logic for loading configuration from a TOML file or using command line arguments.

The Main.java file is also modified to use the CommandLine library and execute the ServerStart command. The previous implementation of starting the server with hardcoded values is replaced with calling the start() method of an instance of the Server class.

Additionally, some changes are made to the Server class:
- The SERVER_ADDRESS and SERVER_PORT fields are moved to be static fields.
- A static initializer block is added to set default values for SERVER_ADDRESS and SERVER_PORT.
- The start() method now throws a RuntimeException if an error occurs when trying to start the server.

These changes allow for more flexibility in configuring and starting the network server.